train/requirements.txt
torch
transformers
trl
datasets
loguru
accelerate
peft
scipy
python-dotenv
openai